task 1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -6,7 +6,7 @@
 *@size: size of the array
 *@action: address of function
 */
-void array_iterator(int *array, size_t size, void (*action)(int))
+void array_iterator(int *array, int size, void (*action)(int))
 {
 	int i;
